adding poll apis and hooks
