+++ client/src/api/index.ts
@@ -1,0 +1,7 @@
+// Poll API exports
+export * from './queries/poll';
+export * from './queries/thread';
+
+// Hook exports
+export * from './hooks/usePoll';
+export * from './hooks/useThreads';
